Secure the api route based on role user

+++ app/Http/Controllers/API/BlogController.php
@@ -6,13 +6,20 @@ use Illuminate\Http\Request;
 use App\Http\Controllers\API\BaseController as BaseController;
 use App\Models\Blog;
 use App\Http\Resources\Blog as BlogResource;
+use Illuminate\Support\Facades\Auth;
 use Illuminate\Support\Facades\Validator;
 
 class BlogController extends BaseController
 {
     public function index()
     {
-        $blogs = Blog::all();
+        $user = auth()->user();
+
+        if ($user->hasRole(['admin','manager'])) {
+            $blogs = Blog::all();
+        } else {
+            $blogs = Blog::where('user_id', $user->id)->get();
+        }
         return $this->sendResponse(BlogResource::collection($blogs), 'Posts fetched.');
     }
 
@@ -21,12 +28,15 @@ class BlogController extends BaseController
         $input = $request->all();
         $validator = Validator::make($input, [
             'title' => 'required',
-            'description' => 'required'
+            'description' => 'required',
+
         ]);
         if($validator->fails()){
             return $this->sendError($validator->errors());
         }
+
         $blog = Blog::create($input);
+
         return $this->sendResponse(new BlogResource($blog), 'Post created.');
     }
 
@@ -36,11 +46,19 @@ class BlogController extends BaseController
         if (is_null($blog)) {
             return $this->sendError('Post does not exist.');
         }
+        if ($blog->user_id !== Auth::id()){
+            return $this->sendError('You dont have access to view the post');
+        }
+
         return $this->sendResponse(new BlogResource($blog), 'Post fetched.');
     }
 
     public function update(Request $request, Blog $blog)
     {
+        if ($blog->user_id !== Auth::id()){
+            return $this->sendError('You dont have access to modify the  post');
+        }
+
         $input = $request->all();
         $validator = Validator::make($input, [
             'title' => 'required',
@@ -58,6 +76,10 @@ class BlogController extends BaseController
 
     public function destroy(Blog $blog)
     {
+        if ($blog->user_id !== Auth::id()){
+            return $this->sendError('You dont have access to delete this post');
+        }
+
         $blog->delete();
         return $this->sendResponse([], 'Post deleted.');
     }
